requirements.txt
fastapi
uvicorn[standart]
sqlalchemy
passlib[bcrypt]
bcrypt
python-jose[cryptography]
pydantic
python-multipart
